Update file structure
Add package(activity, dialog, manager, fragment) and classify classes according to packages.
Change class name: UofSocket -> SocketManager.
