+++ app/src/main/AndroidManifest.xml
@@ -19,43 +19,43 @@
             android:screenOrientation="fullSensor"
             tools:replace="screenOrientation"/>
         <activity
-            android:name=".RegisterActivity"
+            android:name=".activity.RegisterActivity"
             android:exported="true" />
         <activity
-            android:name=".OwnerSettingActivity"
+            android:name=".activity.OwnerSettingActivity"
             android:exported="true" />
         <activity
-            android:name=".OwnerQRRecognitionActivity"
+            android:name=".activity.OwnerQRRecognitionActivity"
             android:exported="true" />
         <activity
-            android:name=".OwnerLobbyActivity"
+            android:name=".activity.OwnerLobbyActivity"
             android:exported="true" />
         <activity
-            android:name=".PayActivity"
+            android:name=".activity.PayActivity"
             android:exported="true" />
         <activity
-            android:name=".OrderingActivity"
+            android:name=".activity.OrderingActivity"
             android:exported="true" />
         <activity
-            android:name=".CardActivity"
+            android:name=".activity.CardActivity"
             android:exported="true" />
         <activity
-            android:name=".SettingActivity"
+            android:name=".activity.SettingActivity"
             android:exported="true" />
         <activity
-            android:name=".OrderListActivity"
+            android:name=".activity.OrderListActivity"
             android:exported="true" />
         <activity
-            android:name=".LobbyActivity"
+            android:name=".activity.LobbyActivity"
             android:exported="true" />
         <activity
-            android:name=".QRRecognitionActivity"
+            android:name=".activity.QRRecognitionActivity"
             android:exported="true" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activity.LoginActivity"
             android:exported="true" />
         <activity
-            android:name=".IntroActivity"
+            android:name=".activity.IntroActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -73,7 +73,7 @@
                     android:scheme="uofmobile" />
             </intent-filter>
         </activity>
-        <activity android:name=".SocketTestActivity" />
+        <activity android:name=".activity.SocketTestActivity" />
     </application>
 
 </manifest>
