+++ app/src/main/java/com/uof/uof_mobile/activity/IntroActivity.java
@@ -1,9 +1,11 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.uof.uof_mobile.R;
 
 public class IntroActivity extends AppCompatActivity {
