+++ app/src/main/java/com/uof/uof_mobile/activity/LobbyActivity.java
@@ -1,4 +1,4 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.widget.Button;
 import android.widget.ImageView;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.uof.uof_mobile.R;
 
 public class LobbyActivity extends AppCompatActivity {
     Button btn_card, btn_orderlist, btn_setting;
