+++ app/src/main/java/com/uof/uof_mobile/activity/OrderingActivity.java
@@ -1,8 +1,10 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.activity;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.uof.uof_mobile.R;
 
 public class OrderingActivity extends AppCompatActivity {
