+++ app/src/main/java/com/uof/uof_mobile/activity/PayActivity.java
@@ -1,14 +1,16 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.activity;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class OwnerSettingActivity extends AppCompatActivity {
+import com.uof.uof_mobile.R;
+
+public class PayActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_ownersetting);
+        setContentView(R.layout.activity_pay);
     }
 }
