+++ app/src/main/java/com/uof/uof_mobile/activity/QRRecognitionActivity.java
@@ -1,4 +1,4 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.zxing.integration.android.IntentIntegrator;
 import com.google.zxing.integration.android.IntentResult;
+import com.uof.uof_mobile.R;
 
 public class QRRecognitionActivity extends AppCompatActivity {
     private IntentIntegrator QRscan;
