+++ app/src/main/java/com/uof/uof_mobile/dialog/RegisterTypeDialog.java
@@ -1,4 +1,4 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.dialog;
 
 import android.app.Dialog;
 import android.content.Context;
@@ -8,6 +8,8 @@ import android.view.WindowManager;
 import androidx.annotation.NonNull;
 import androidx.appcompat.widget.AppCompatButton;
 import androidx.appcompat.widget.AppCompatImageButton;
+
+import com.uof.uof_mobile.R;
 
 public class RegisterTypeDialog extends Dialog {
     private final Context context;
