+++ app/src/main/java/com/uof/uof_mobile/fragment/SettingPreferenceFragment.java
@@ -1,4 +1,4 @@
-package com.uof.uof_mobile;
+package com.uof.uof_mobile.fragment;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -8,6 +8,8 @@ import android.preference.Preference;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.widget.Toast;
+
+import com.uof.uof_mobile.R;
 
 
 public class SettingPreferenceFragment extends PreferenceFragment {
