+++ app/src/main/res/layout/activity_intro.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IntroActivity">
+    tools:context=".activity.IntroActivity">
 
     <TextView
         android:id="@+id/textView3"
