+++ app/src/main/res/layout/activity_ordering.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OrderingActivity">
+    tools:context=".activity.OrderingActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
