+++ app/src/main/res/layout/activity_orderlist.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OrderListActivity">
+    tools:context=".activity.OrderListActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
