+++ app/src/main/res/layout/activity_ownerlobby.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OwnerLobbyActivity">
+    tools:context=".activity.OwnerLobbyActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
