+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".RegisterActivity">
+    tools:context=".activity.RegisterActivity">
 
     <!-- 일반유저 회원가입 -->
     <androidx.appcompat.widget.LinearLayoutCompat
