+++ app/src/main/res/layout/activity_setting.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/settings_fragment"
-    android:name="com.uof.uof_mobile.SettingPreferenceFragment"
+    android:name="com.uof.uof_mobile.fragment.SettingPreferenceFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     />
