+++ app/src/main/res/layout/activity_sockettest.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SocketTestActivity">
+    tools:context=".activity.SocketTestActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/til_sockettest_serverip"
